add a feature to show the PIDs of processes optionally
git-svn-id: https://powertop.googlecode.com/svn/trunk@273 0bbd23ad-e62f-0410-b3b3-4ddab988b7fa

+++ process.c
@@ -79,7 +79,7 @@ void suggest_process_death(char *process_match, char *tokill, struct line *sline
 	int i;
 
 	for (i = 0; i < linecount; i++) {
-		if (strstr(slines[i].string, process_match)) {
+		if (slines[i].string && strstr(slines[i].string, process_match)) {
 			char hotkey_string[300];
 			sprintf(hotkey_string, _(" K - kill %s "), tokill);
 			strcpy(process_to_kill, tokill);
